Rename util to utilities.
