+++ safe/gui/tools/wizard/step_fc40_explayer_from_browser.py
@@ -8,7 +8,7 @@ from safe.definitions.layer_purposes import layer_purpose_exposure
 from safe.utilities.resources import resources_path
 from safe.gui.tools.wizard.wizard_step import get_wizard_step_ui_class
 from safe.gui.tools.wizard.wizard_step_browser import WizardStepBrowser
-from safe.gui.tools.wizard.wizard_utils import layers_intersect
+from safe.gui.tools.wizard.utilities import layers_intersect
 
 from safe import messaging as m
 from safe.utilities.i18n import tr
