+++ safe/gui/tools/wizard/step_fc55_agglayer_from_canvas.py
@@ -15,7 +15,7 @@ from safe.utilities.i18n import tr
 
 from safe.gui.tools.wizard.wizard_step import get_wizard_step_ui_class
 from safe.gui.tools.wizard.wizard_step import WizardStep
-from safe.gui.tools.wizard.wizard_utils import layers_intersect
+from safe.gui.tools.wizard.utilities import layers_intersect
 
 __copyright__ = "Copyright 2016, The InaSAFE Project"
 __license__ = "GPL version 3"
