+++ safe/gui/tools/wizard/step_kw47_default_inasafe_fields.py
@@ -19,7 +19,7 @@ from safe.definitions.layer_purposes import (layer_purpose_aggregation)
 from safe.definitions.utilities import get_fields, get_compulsory_fields
 from safe.definitions.layer_geometry import layer_geometry_raster
 from safe.definitions.constants import no_field
-from safe.gui.tools.wizard.wizard_utils import get_inasafe_default_value_fields
+from safe.gui.tools.wizard.utilities import get_inasafe_default_value_fields
 
 from safe.gui.tools.wizard.wizard_step import (
     WizardStep, get_wizard_step_ui_class)
