+++ safe/gui/tools/wizard/step_kw49_inasafe_raster_default_values.py
@@ -16,7 +16,7 @@ from safe_extras.parameters.qt_widgets.parameter_container import (
 
 from safe.definitions.layer_purposes import (layer_purpose_aggregation)
 from safe.definitions.utilities import get_fields, get_compulsory_fields
-from safe.gui.tools.wizard.wizard_utils import get_inasafe_default_value_fields
+from safe.gui.tools.wizard.utilities import get_inasafe_default_value_fields
 
 from safe.gui.tools.wizard.wizard_step import (
     WizardStep, get_wizard_step_ui_class)
