+++ safe/metadata/generic_layer_metadata.py
@@ -19,7 +19,7 @@ __copyright__ = ('Copyright 2012, Australia Indonesia Facility for '
 
 from xml.etree import ElementTree
 from safe.metadata import BaseMetadata
-from safe.metadata.utils import reading_ancillary_files, prettify_xml
+from safe.metadata.utilities import reading_ancillary_files, prettify_xml
 
 
 class GenericLayerMetadata(BaseMetadata):
