+++ safe/metadata/output_layer_metadata.py
@@ -6,9 +6,9 @@ from xml.etree import ElementTree
 from safe.metadata import BaseMetadata
 
 from safe.metadata.provenance import Provenance
-from safe.metadata.utils import reading_ancillary_files, prettify_xml
-from safe.metadata.utils import XML_NS
-from safe.metadata.utils import merge_dictionaries
+from safe.metadata.utilities import reading_ancillary_files, prettify_xml
+from safe.metadata.utilities import XML_NS
+from safe.metadata.utilities import merge_dictionaries
 from safe.metadata.encoder import MetadataEncoder
 
 __copyright__ = "Copyright 2016, The InaSAFE Project"
