+++ safe/metadata/utilities.py
